requirements.txt
annotated-types==0.7.0
anyio==4.9.0
argon2==0.1.10
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
contourpy==1.3.2
cryptography==45.0.5
cycler==0.12.1
ecdsa==0.19.1
fastapi==0.115.14
fonttools==4.58.4
greenlet==3.2.3
h11==0.16.0
huepy==1.2.1
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
kiwisolver==1.4.8
livereload==2.7.1
markdown2==2.5.3
MarkupSafe==3.0.2
matplotlib==3.10.3
nodeenv==1.9.1
npm==0.1.1
numpy==2.3.1
optional-django==0.1.0
packaging==25.0
pandas==2.3.0
pandas-stubs==2.2.3.250527
pillow==11.2.1
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyMySQL==1.1.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-jose==3.5.0
pytz==2025.2
requests==2.32.4
rsa==4.9.1
scikit-learn==1.7.0
scipy==1.16.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
threadpoolctl==3.6.0
timeframe==3.0.3
tornado==6.5.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.34.3
vite==1.5.2
wheel==0.45.1
